Move SuspenseProps to ReactTypes
And add some annotations
