+++ packages/react-reconciler/src/ReactFiberSuspenseComponent.js
@@ -21,18 +21,6 @@ import {
   isSuspenseInstanceFallback,
 } from './ReactFiberConfig';
 
-export type SuspenseProps = {
-  children?: ReactNodeList,
-  fallback?: ReactNodeList,
-
-  // TODO: Add "unstable_" prefix?
-  suspenseCallback?: (Set<Wakeable> | null) => mixed,
-
-  unstable_avoidThisFallback?: boolean,
-  unstable_expectedLoadTime?: number,
-  unstable_name?: string,
-};
-
 // A null SuspenseState represents an unsuspended normal Suspense boundary.
 // A non-null SuspenseState means that it is blocked for one reason or another.
 // - A non-null dehydrated field means it's blocked pending hydration.
